refactor: more registries to component sets, eliminate an unneeded one
